Allow transaction slots to be reused after transaction is finished.
To reuse the slot, it needs to ensure that (a) the xid is committed
and all-visible (b) if the xid is committed, then write an undo
records for all the tuples that belong to that slot.  The undo record
will cover the transaction information of committed transaction. The
basic idea is that after reuse of slot, if someone still needs to find
the information of prior transaction that has modified the tuple, the
same can be retrieved from undo.

We can reuse the slot even after the transaction has aborted, but for
that we need to have Rollback facility which is currently not
implemented, so we can't reuse slots of aborted transactions.

Amit Kapila and Dilip Kumar.