Write WAL for invalidating the transaction slot
The important consideration for logging invalidating slot operation is
that we need to identify the tuples pointing to these slots and also the
transaction id in the slot while performing redo operation for undo log.
So, we are conditionally (when full page writes are off) writing tuple
offsets and slot xid in WAL as well.  In case full page writes are
on, we can rely on the tuples and slots from page.

Path by Dilip Kumar, Reviewed by Kuntal Ghosh and Amit Kapila