Support Insert .. On Conflict
The design is similar to current heap such that we use the
speculative token to detect conflicts.  We store the speculative token
in undo instead of in the tuple header (CTID) simply because zheap’s
tuple header doesn’t have CTID. Additionally, we set a bit in tuple
header to indicate speculative insertion.  ZheapTupleSatisfiesDirty
routine checks this bit and fetches a speculative token from undo.

Amit Kapila and Ashutosh Sharma