Introducing a queue for passing rollback requests to undo worker
To increase the efficiency of rollback mechanism in zheap,
we now have a rollback queue. Now, any rollback request that
exceeds the threshold -- rollback_overflow_size, is added to
this queue. Whenever undo worker is idle it checks if rollback
queue has some entries and executes the required undo actions,
if any. The rollbacks required for 'rollbacks to savepoint' are
not added to this queue, rather the backend itself executes the
required undo actions for them.

The rollback queue is implemented as a shared hash table.

Reviewd by Beena Emerson and Amit Kapila