Implement Foreign Key Constraint for zheap
For zheap, when we call the triggers we convert a zheap tuple to heap
tuple. But, we don't store any transaction related information on the
heap tuple. Hence, I've to make the following two
assumptions/considerations for the foreign key implementation in
zheap:

1. RI_FKey_fk_upd_check_required: In this function, we check if the
original row was inserted by our own transaction, we must fire the
trigger whether or not the keys are equal. For zheap, since we don't
have the transaction related information, we always fire the trigger.
So, even if the keys are equal, we cannot skip the trigger for zheap.

2. validateForeignKeyConstraint: During ALTER TABLE..ADD CONSTRAINT
FOREIGN KEY, this function is called to validate whether we can add
the foreign key constraint. First, it tries to fire a LEFT JOIN query
to test the validity. If the user doesn't have proper access rights to
pktable/fktable, we scan the table in pagescanmode and fire trigger
for each tuple. Before firing the trigger, it takes buffer lock to
check whether it can skip the trigger. For zheap, we don't retain the
pin on the buffer. Hence, we've to read the buffer again before
locking the same.

Patch by me, Reviewed and tested by Ashutosh Sharma