Avoid calling PageGetUNDO
We need to avoid calling PageGetUndo as it re-access the transaction
slots the second time.  We can already get it via PageReserveTransactionSlot.
This is okay till now, but with TPD, we need to again access the TPD page
which will be costly.

Patch by me, reviewed and edited by Dilip Kumar