Fix bugs related to concurrency in zheap update
While updating the tuple we might need to release the buffer lock and
reaquire same. During that period where we do not hold buffer lock
a concurrent process might have moved the tuple to undo and/or pruned
the page. So whenever we reacquire the buffer lock check if
itemid is deleted and readjust the position of tuple in page buffer.

Patch by Mithun C Y and Amit Kapila.