Prevent usage of uninitialized tuple during tuple routing.
TransitionCaptureState.tcs_original_insert_tuple is of type HeapTuple,
and it was assigned a tuple variable which remains unitinitialized in
case of zheap table. Fix it so that the zheap tuple is converted to
heap tuple and then assigned to tcs_original_insert_tuple. Fixed this
in both ExecPrepareTupleRouting() and CopyFrom(). Due to this issue,
trigger.sql regression test used to crash on some environments.

Amit Khandekar, reviewed by Kuntal Ghosh.