Handling when a transaction span across undo logs and avoid single WAL logged operation to span across multiple log.
The first part of the patch handle a case, when a single WAL logged
operation which needs multiple undo records (e.g non-inplace update,
multi-insert) we avoid it to go in multiple logs.  For that we first
allocate all the undo required for the operation in one allocate call.

And the second part handle the discarding and rollback when a transaction
span across undo logs.

Path by Dilip Kumar Reviewed by Amit Kapila.