Handle aborted xact in ZHeapTupleSatisfiesOldestXmin
If the latest transaction for the tuple aborted, we fetch a prior committed
version of the tuple and return it along with prior comitted xid and status
as HEAPTUPLE_LIVE.
If the latest transaction for the tuple aborted and it also inserted
the tuple, we return the aborted transaction id and status as
HEAPTUPLE_DEAD. In this case, the caller *should* never mark the
corresponding item id as dead. Because, when undo action for the same
will be performed, we need the item pointer.

Patch by Amit Kapila and me. Reviewed by Amit Kapila.