Handle locked-only tuple in ZHeapTupleSatisfiesOldestXmin
In ZHeapTupleSatisfiesOldestXmin, we can't take any decision if the
tuple is marked as locked-only. It's possible that inserted transaction
took a lock on the tuple. Later, if it rolled back, we should return
HEAPTUPLE_DEAD, or if it's still in progress, we should return
HEAPTUPLE_INSERT_IN_PROGRESS. Similarly, if the inserted transaction
got committed, we should return HEAPTUPLE_LIVE. The subsequent checks
in the function already takes care of all these possible scenarios,
so we don't need any extra checks for locked-only tuple.