Fix silly bug in undolog_xlog_discard().
When an XLOG_UNDOLOG_DISCARD record is replayed, we need to tell the
checkpointer to forget about any files that we are about to unlink.
I was using the wrong variable, so if a single XLOG_UNDOLOG_DISCARD
record caused segment files 1, 2, 3 to be unlinked, I was telling it
to forget about fsyncing 3, 3, 3.  Then it would eventually try to
fsync 1 and 2 and try to raise an error.  Repair that.

Additionally, in the error path resulting from the above bug, I was
also calling FilePathName() on a File that I had failed to open.  That
causes an assertion failure.  Repair that too.

Thomas Munro, reported by Neha Sharma, RM43571