The startup process shouldn't attach to undo logs.
When replaying XLOG_UNDOLOG_META records, the startup process was
recording that it was attached to the referenced undo log.  That
caused corruption of the freelists when it tried to detach on exit.
During recovery we shouldn't attach at all; instead we use the
xid->undo log mapping.

Thomas Munro, RM43614