An optimization to improve COPY FREEZE in zheap
In COPY FREEZE, when we insert a tuple, we always mark it as frozen.
Hence, there is a possible performance optimization for the same scenario:
1. We can skip inserting undo records for the tuples to be inserted.
2. There is no need to reserve a transaction slot.

Here is the implementation details:
1. Set skip_undo = true if HEAP_INSERT_FROZEN is mentioned.
2. If skip_undo is true, we don't have to reserve a transaction slot in
the page. Also, we skip preparing and inserting undo records for the
to-be-inserted tuples.
3. For recovery, a new WAL flag XLZ_INSERT_IS_FROZEN is introduced. It's
true if HEAP_INSERT_FROZEN is mentioned. During WAL replay, we skip preparing
and inserting undo records if XLZ_INSERT_IS_FROZEN is set in WAL records.

Patch by me. Reviewed by Amit Kapila.