Store Itemids in TPD page.
Earlier heappage always point directly to the actual offset of TPD
entry in a TPD page.  This won't work after pruning as even if the
particular page's TPD entry is not pruned, we might not be able to
directly access the TPD entry as offset might have moved. Now, we can
reach our TPD entry if we can traverse all the TPD entries and TPD
entry has block number in it, but that is quite inefficient.  To
overcome this problem, it is better to store Itemids in the TPD page.

Amit Kapila, reviewed and verified by Dilip Kumar