Report the scan location for zheap meta page
During scan, we report our scan position for synchronization
purposes. We do this before checking for end of scan so that the
final state of the position hint is back at the start of the
rel. But, if we skip metapage, the scan location may point to a
block at the end of the relation.

Patch by me, reviewed by Amit Kapila