An optimization for temp tables
For temp relations, we don't have to check all the slots since
no other backend can access the same relation. If a slot is available,
we return it from here. Else, we freeze the slot in PageFreezeTransSlots.

Note that for temp tables, oldestXidWithEpochHavingUndo is not relevant as
the undo for them can be discarded on commit.  Hence, comparing xid
with oldestXidWithEpochHavingUndo during visibility checks can lead to
incorrect behavior.  To avoid that, we mark the tuple as frozen for any
previous transaction id.  In that way, we don't have to compare the previous
xid of tuple with oldestXidWithEpochHavingUndo.

Patch by me. Reviewed by Amit Kapila.