Allow inplace updates if the row can fit on the same page.
Currently, we perform inplace updates only when the new row is smaller than
old row or if the old row is the last row on page and it has space after
it. However, there are more cases where we can perform inplace updates:
(a) If there's no free space immediately following the tuple, but there is
a space in the page to accomodate the entire tuple. (b) If there's no free
space immediately following the tuple, but there is a space in page to
accommodate the delta tuple (new_tuple_size - old_tuple_size).

We allow pruning function to rearrange the page such that it can make space
adjacent to the tuple being updated. This is only possible if the page has
at least space to support equal to (newtupsize - oldtupsize). Otherwise,
also we try to prune the dead/deleted tuples to see if the new tuple can be
accommodated on same page and that will allow inplace updates.

To perform pruning, we make the copy of the page. We don't scribble on
that copy, rather it is only used during repair fragmentation to copy the
tuples. So, we need to ensure that after making the copy, we operate on
tuples, otherwise, the temporary copy will become useless. It is okay
to scribble on itemid's or special space of page.  While rearranging the
page tuples will be placed in itemid order. It will help in the speedup of
future sequential scans. Note that we use the temporary copy of the page to
copy the tuples as writing in itemid order will overwrite some tuples. We have
also changed the patch such that REDO will perform repairfragmentation only if
we it has been done during DO operation.

Amit Kapila, Mithun C Y.