Introducing sub-transactions lock type in zheap
With this patch, the sub-transactions will have a new type
of lock for them. Now, instead of waiting for top-level
transaction lock, the tuples are free once the sub-transaction
is committed. This way, we neither waste our transaction-slots
by assigning transaction ids to sub-transaction, nor we suffer
in performance by waiting on top-level transactions.

Patch by Amit Kapila and me.