Support Cluster/Vacuum Full in zheap.
As of now we only rewrite LIVE tuples and we freeze them before storing in new
heap.  This is not a good idea as we lose all the visibility information of
tuples, but OTOH, the same can't be copied from the original tuple as that is
maintained in undo and we don't have facility to modify undorecords.  We
have some ideas how to do that and those are documented in rewritezheap.c.

Patch by Amit Kapila with help from Mithun C Y who also reviewed the patch.