Fix tuple handling with trigger functions and transition capture.
In ExecUpdate(), for zheap tables, ExecARUpdateTriggers() was called
only if resultRelInfo->ri_TrigDesc is true, which is incorrect. This
function also needs to be called for transition capture even when
trigdesc is false. Due to this, transition table rows were not getting
generated.

Furthermore, in ExecInsert(), for zheap tables, ExecARUpdateTriggers()
was getting called using tuple when it should be called using ztuple.

Fixed both these by making the newtuple parameter of
ExecARUpdateTriggers() a void * type and renaming it to
newtuple_abstract, so that we can pass tuple or ztuple according to
the table storage. And then in ExecARUpdateTriggers(), do the
conversion from zheap to heap if it's a zheap tuple.

On similar lines, do changes for ExecARInsertTriggers(), where
trigtuple is made trigtuple_abstract. Now that this function accepts
an abstract tuple, fix another pending issue in copy.c : in CopyFrom()
and CopyFromInsertBatch(), pass either zheap or heap tuples to
ExecARInsertTriggers().

Reviewed by Dilip Kumar and Amit Kapila.