Support inplace update of TPD entries.
Whenever TPD entry needs to be extended, we call Call TPDPagePrune to
ensure that it will create a space adjacent to current offset for the new
(bigger) TPD entry, if possible. We use compactify_ztuples to ensure that
space adjacent to existing entry can be created. For inplace update of TPD
entry, we just replace the old entry with new entry at the same location
similar to inplace updates of tuples. We also write WAL record this
operation.

Amit Kapila and Kuntal Ghosh, reviewed and verified by Dilip Kumar