Add expected file for eval-plan-qual isolation test
Include metapages in ctid for zheap tuples. Also, Updated a test case
related to self join. Basically, when performing a self join if it needs
to pass through EvalPlanQualFetch path, it's possible that both sides of
the join see the same value due to in-place update. This behaviour is
different from heap, but similar to other undo-based storage.