Return error on trying to update a row moved to another partition.
The approach used here is to set the ZHEAP_MOVED (ZHEAP_DELETED |
ZHEAP_UPDATED) flag when the tuple is moved to different partition.
All the cases that are handled in heap needs to be handled for zheap as
well.

Amit Kapila, based on earlier patch by Amit Khandekar which has used a
different approach.