Add expected file for stats.sql regression test
In zheap, we increase pgstat_info->trans->tuples_updated only for
non-inplace updates, otherwise vacuum will be triggered for in-place
updates as well. But for heap, we always increase tuples_updated since
heap always creates a new version of the tuple during updates. Hence, we
need to fix the output in expected file for zheap.