Add expected file for rowsecurity regression test
In rowsecurity test, there is a test case that uses table sample scan
with bernoulli distribution. The output of tablesample scan partially
depends on the block number of the relpage from which the tuples are
fetched. For zheap, blocknumber 0 is meta page. Hence, tuples are stored
from block 1. Hence, the output of table sample scan can be different.