Add empty TPD pages to FSM.
The empty TPD pages are added to FSM by vacuum which prune such pages as
well. The empty pages from FSM can be used either by zheap or TPD when
required. We need to ensure that when we access TPD pages, such a page can
be pruned or truncated away by vacuum. After pruning the TPD page, it can
be freed by removing it from the chain of TPD pages.

As of now, only vacumm can add TPD pages into FSM, but we might want to
improve it someday that backends can also do the same.

Patch by me with help from Dilip Kumar who also tested and verified the
patch.