Fix a bug for projecting table oid
There are certain places in the code where we form a heap tuple using
ExecCopyTuple and store it in slot along with the zheap tuple. In that
case, we don't copy the tableOid to heap tuple. Hence, for fetching
tableOid, we should prefer a zheap tuple instead of a heap tuple.