Prune entire TPD page at one-shot if possible.
We have used tpd_latest_xid_epoch stored in page to prune the entire TPD page.
Basically, if tpd_latest_xid_epoch precedes oldestXidhaving undo, then we
can assume all the entries in the page can be pruned.

Apart from that, I have changed the logic so that page is freed during
pruning if all the entries are removed from the page. this will ensure that
pages will be reclaimed whenever they are empty, not only during vacuum.

In the passing, I have fixed another related issue which is after we get
the new page from fsm, we need to ensure that it is a tpd page before
using it.

Patch by Amit Kapila, reviewed and verified by Dilip Kumar