Remove relfilenode and tablespace id from the undo record and store reloid
There was no reason why we stored relfilenode and tablespaceid in undo record,
instead we could just store reloid.  Earlier, we might have kept it thinking
that we will perform rollback without database connection but that's not the
case now.  This will save the space as well as it will be useful when we need
to transfer the undo records across the relfilenodes e.g ALTER TABLE SET TABLESPACE.

Dilip Kumar reviewed by Amit Kapila