Fix the modification of TPD map during rollback.
If the previous transaction slot points to a TPD slot then we need to
update the slot in the offset map of the TPD entry.  This is the case where
during DO operation the previous updater belongs to a non-TPD slot whereas
now the same slot has become a TPD slot.  In such cases, we need to update
offset-map.

Patch Mahendra Singh, reviewed by Amit Kapila