Fix issues in TPD
1. If previous and next block number is invalid for a TPD page, then
that is the only TPD page in the same relation. We should handle the
case correctly in TPDFreePage.
2. While allocating a new TPD page, we ask a new page from FSM. It's
possible that FSM returns a zheap page on which the current backend already
holds a lock in exclusive mode. Hence, try for a conditional lock. If it
can't get the lock immediately, extend the relation and allocate a new
TPD block.
3. In TPDPageLock, if a TPD buffer is already pruned, we don't take a
lock on the same. Here, we should take the opportunity to clear the TPD
location from the corresponding zheap buffer.
4. In lazy vacuum, we don't vacuum a TPD page. Instead, we try to prune
the page. But, if it's already pruned, we should skip it.
5. There are several places in the code where we lock a TPD page before
entering critical section. For non-inplace updates on different buffers,
it is possible that both old and new zheap buffer corresponding to the
same TPD buffer. Hence, we should be careful that we don't try to lock
the TPD page two times. Else, we'll be waiting on self.

Patch by me. Reviewed by Amit Kapila.