Teach undolog.c to track transactions itself.
Track the location of the current and previous transaction and previous
record length automatically in the undo log meta-data.

In recovery, remove the xid->undo log mapping code that is now redundant.

Remove the remaining traces of XLOG_UNDOLOG_ATTACH and XLOG_UNDOLOG_META
WAL records, which are now entirely redundant.

Thomas Munro