Fix buffer lock in RelationGetBufferForZTuple
Whenever we need two buffers for updating a tuple (non-inplace),
we use the rule "lock lower numbered buffer first" to avoid
deadlocks. But, in zheap this is not the sufficient condition.
It's possible that the new buffer is a pruned TPD buffer and some
other backend is trying to use it while holding lock on a zheap
buffer with higher block number. To fix this, whenever we encounter
a TPD buffer as a new buffer for updates and it has lower block
number,we allocate a new block by extending the relation.

Patch by Amit Kapila, tested by Mahendra Singh Thalor.