Fix conditions to check whether tpd entry exists in TPD page
To check whether a TPD entry exists, we check whether the page is empty
and it is indeed a TPD page. Apart from that, we should also check
whether the item offset for the tpd entry exists in the page. If it
exists, then only we can check whether the corresponding tpd entry
matches the required block number.

Patch by me. Reviewed by Amit Kapila.