Design change for handling multi-log transaction in undo.
In current patch I have removed the concept of prevlogno in undo log meta.
For addressing these issues related to multilog I have changed the
design as below
- Now, at 'DO' time we identify the log switch by identifying which
log we are attached to before and after allocate. And, if the log is
switched we write a WAL for the same and during recovery whenever
this WAL is replayed we stored the undo record pointer of the
transaction header (which is in the previous undo log) in
UndoLogStateData and read it while allocating space the undo record
and immediately reset it.

- For handling the discard issue, along with updating the current
transaction's start header in the previous undo log we also update the
previous transaction's start header in the current log if we get
assigned with the non-empty undo log.

- For, identifying the previous undo record of the transaction during
rollback (when undo log is switched), we store the transaction's last
record's (in previous undo log) undo record pointer in the transaction
header of the first undo record in the new undo log.

Dilip Kumar reviewed by Amit Kapila