Fix a assert failure during recovery
The ItemIdSetUnusedExtended marks itemIds as UNUSED with transaction
slot information. However, the ItemIdUnused is called later by the
ZPageRepairFragmentation resetting this transaction slot information
causing assert failure in recovery to get slot from TPD because in
recovery, we are not setting slot for unused xlog.
If unused_set flag is true, then we will not reset slot info so
ignoring fetching slot info based on this flag. Because anyway if
unused_set is set, then we will not re-evaluate.

Patch by Mahendra Singh Thalor, reviewed by Amit Kapila