Fix tuple slot for LockForUpdate
When we've to lock the tuple before updating it, we take LockForUpdate
lock on the tuple. In that case, we also set the current transaction
slot on the tuple. This avoids the requirement of marking the tuple with
multilocker flag when previous updater/inserter is not all-visible. But
when
the tuple already has a multilocker flag, we should maintain the old
transaction slot.
In the above case, when the tuple has to be marked with multilocker
flag, we should propagate the updaters information as well. Otherwise
the tuple is marked with locked-only flag and current transaction slot
is set. So, there is no need to propagate the updaters information.
This patch also fixes the previous tuple xid that'll be stored in undo
record for ZHEAP_UPDATE.

Patch by Amit Kapila and Kuntal Ghosh