Bug fix in undo apply worker
Currently, undo launcher is acuiring a lock on database before launching
undo worker and that creates deadlock if concurrently backend tries to
drop the database.  Now, for handling this, launcher is just checking whether
dbid is exist or not if it does not exist it will remove all undo actions
for the db from the rollback hash table.

Rafia Sabih reviewed by Dilip Kumar and Amit Kapila