Fix a buffer locking issue in zheap_lock_updated_tuple
In zheap_lock_updated_tuple, when we lock the tuple, we've to reserve
one transaction slot in the page. During page reserve, if we've
reaquired the lock, we've to check the tuple status again. In that case
we should unlock the buffer before locking it again.

Patch by me. Reviewed by Amit Kapila.