Prevent deadlocks while acquiring lock on TPD page.
To prevent deadlock risks, we always need to ensure that lower
numbered TPD block is locked first. This is generally followed
in all the places in code, but we missed to do it while acquiring
lock on TPD page during non-in-place update.

Patch by Mahendra Singh Thalor, reviewed and edited by Amit Kapila