Avoid fetching transaction information from undo in some cases
We can avoid fetching the transaction information from the undo even
if the the slot is reused by another transaction if the slot's current
xid is all-visible, then the xid prior to it must be all visible.
The other case where we can avoid it when the current xid is visible
to the snapshot for similar reasoning.

Dilip Kumar and Kuntal Ghosh reviewed by Amit Kapila