In ZHeapPageGetCtid, handle multilocker undo records
When a tuple is updated, it can be locked in key share mode by a
concurrent locker. If the locker reserves the same slot id as the
updater's xid, the latest undo pointer in the slot will refer to
the locker's undo record. In that case, we should fetch the prior undo
record to get the updater's info.

Patch by me. Reviewed by Amit Kapila.