pg_stats: Fix the dead tuple count
Update counters were revised to store counts of both inplace and
non-inplace updates in 122eef23ca but the dead tuple calculation
was not adjusted to exclude inplace updates which do not generate
dead tuples and this triggered vacuum frequently.

This commit adds new counters for inplace updates and modifies the
calculation of dead tuples to:
     dead tuples = updates - inplace updates + delete

Reusing the hot_update counter does not work because calculation of
dead tuples is done at transaction level and hot_update
counter is a non-transactional table level counter.

Patch by me, reviewed by Amit Kapila