Fix "could not read block" error
In ExtendTPDEntry, we are calling TPDPagePrune with can_free as
false.  If TPDPagePrune make TPD page fully empty then we are not
removing that page from meta list to avoid deadlock. And we are not
extending our TPD entry in that page due to empty page.
At vacuum time, we are removing only non-empty pages from meta list.
But it is possible that empty TPD page yet not removed from meta list
and we truncated that page.
To allocate new TPD page, first we check that last used TPD page
have enough space or not.  If that empty TPD page is the last used
TPD page, then we can't read that page because page is already truncated
but not removed from meta list.
So fixed this problem by removing empty TPD page from meta list at
vacuum time if yet not removed.

Patch by Mahendra Singh Thalor, reviewed by Amit Kapila