Fix assert of TPD slot
In zheap_search_buffer, we were reading tuple from undo and trying
to lock tid.
To fetch the xmin (aka transaction that has inserted the tuple), we
need to use the transaction slot of the tuple in the page instead of
the tuple from undo, otherwise, it might traverse the wrong chain.

Patch by Mahendra Singh Thalor, reviewed by Amit Kapila