Bug fix in IndexBuildZHeapRangeScan
The tuple was not freed during the scan and hence the memory
was accumulating, causing an OOM error.

This commit converts the tuple scan code in IndexBuildZHeapRangeScan
to slot API and also sets the sholdFree flag in ExecStoreZTuple to
true when scan type is not rs_pageatatime.

patch by me, review by Kuntal Ghosh and Amit Kapila