Return a struct from ZHeapTupleGetTransInfo.
Large numbers of individual out parameters are difficult to
maintain. For example, adding a new one requires updating many
call sites; and renaming one risks inconsistencies between the
parameter names and the variable names used in the caller.

Use the same struct in ZHeapTupleSatisfiesUpdate.  Callers don't
currently need all of the information contained in the structure,
but since we are retrieving it anyway, there's no harm in
returning the whole thing.

The case of calling ZHeapTupleGetTransInfo solely to fetch the
XID that last touched the tuple seems to be quite common, so
provide a convenience function that does that and use it wherever
it makes sense.