Introduce ZVersionSelector to reduce code duplication.
The ZHeapTupleSatisfies* functions in zheapam_visbility.c basically
all do one of three things: return NULL, return the tuple passed
to them, or hand the request off to GetTupleFromUndo().  Instead
of repeating the GetTupleFromUndo() call in many places, set an
enum variable to indicate which of those three things we're going
to do, and then take the appropriate action at the end of the
function based on the value of the enum.

I think there's more that can be done here to further reduce
code duplication and organize things more logically, but even
doing this much helps a little.