Refactor ZHeapTupleSatisfiesMVCC/ZHeapTupleSatisfiesSelf.
As it turns out, the ZHeapSelectVersion function introduced in
commit fb8cd5c33a0117730cdf3b1af76066a2a1480a19 implements precisely
the same set of rules as ZHeapTupleSatisfiesMVCC in the case of an
MVCC snapshot, and precisely the same set of rules as
ZHeapTupleSatisfiesSelf for a non-MVCC snapshot.

To avoid code duplication, this commit consolidates three sets of
visibility rules into two.  Before this, we had one copy of the
SnapshotMVCC rules, one copy of the SnapshotSelf rules, and a third
set of rules that was a combination of the two.  Now the combined copy
is gone, so we just have one copy of each set of rules.

It is unclear why the places that now rely on ZHeapSelectVersionMVCC
and ZHeapSelectVersionSelf -- and which before today had their own
copies of these rules -- treat all non-MVCC snapshots in the same
manner.  Add some comments encouraging further study of that
question.