Replace MakeEpochXid macro.
This macro only works if the epoch is passed as a uint64, but the
normal type for an epoch is uint32, which led to code contortions
in a number of places.  Replace this macro with an inline function,
called FullTransactionIdFromEpochAndXid for Thomas Munro's recent
work on the PostgreSQL master branch.

Rip out most of the places where epochs are stored as uint64 rather
than uint32, but leave FetchTransInfoFromUndo alone for the time
being, since it needs more extensive surgery.