Assorted code cleanup in zheapam_visibility.c.
In FetchTransInfoFromUndo, xid cannot be NULL, because we dereference
it unconditionally.  So there's no point in checking whether it's
NULL later in the function, but let's Assert() that it isn't.

In ZHeapPageGetNewCtid, make a bit more extensive use of
ZHeapTupleTransInfo and clean up epoch handling.  Actually, setting
the value of zinfo.epoch_xid here is useless, but it's also cheap, and
we eventually need to move to using 64-bit XIDs in a lot more places,
so this is a step in that direction.

Also clean up epoch handling in ZHeapTupleGetTransInfo,
GetVisibleTupleIfAny and ZHeapGetVisibleTuple.