Remove GetTupleFromUndoWithOffset and GetVisibleTupleIfAny.
These functions duplicate GetTupleFromUndo almost exactly.  The major
difference is that GetTupleFromUndo does everything in one function,
while in the other case the exact same logic is split across two
functions.  I'm not sure whether it's better in one function or two,
but I am sure that it's better not to have two copies of the logic.