Modify zheap_tuple_attr_equals
To prevent frequent calls to zheap_deform_tuple, modify
zheap_tuple_attr_equals to handle a Bitmapset of attributes instead
of a single attribute number. It will iterate over the given
attribute list and return all the columns that are different between
the two tuples under consideration. Also, add a new parameter
att_count to zheap_deform_tuple which specifies the number of
attributes to be deformed for the given tuple. Remove znocachegetattr,
zfastgetattr and zheap_getattr that are no longer required.