Remove unnecessary ReadBuffer call in zheap_getsysattr
Though commit 14fdcc4b9 removed support for xmax, cmin, and cmax,
the condition in zheap_getsysattr to read valid buffer was unaltered.
This caused uncessary fetching of buffer when not required. Since only
xmin requires a valid buffer, modify and shift the code to read the
buffer to the appropriate switch case block.