Move 'nobuflock' handling into ZHeapTupleGetTransXID.
None of the direct callers of ZHeapTupleGetTransInfo require
nobuflock = true, so move that logic into the wrapper function,
ZHeapTupleGetTransXID, which has a few callers that do pass
nobuflock = true.

This involves a small amount of duplicated work but it seems
worth it to simplify ZHeapTupleGetTransInfo.  I suspect that
ZHeapTupleGetTransXID is going to get refactored out of
existence eventually anyway.