Bug fix in undo record bulk fetch
During bulk fetch if we have already read the to_urecptr and after that
the memory segment is full then we need not to call BulkFetch again,
but due to this bug it will call the BulkFetch and it will start fetching
from the previous urec pointer of the to_urecptr so it will never reach to
end until it will find prevlen as 0.  So, it will fetch many unwanted or
already discarded undo records which doen't belongs to this transaction.

Dilip Kumar