Move CID visibility checking to a separate function, ZHeapCheckCID.
Moving this code out of ZHeapSelectVersionMVCC makes it easier to
postpone fetching the CID until it's clear whether we actually need
it.  In this commit, GetTupleFromUndo is optimized to work that way.
ZHeapTupleSatisfies needs a similar optimization, but the current
code structure makes that hard to implement, so I'm leaving that
for a later commit.

Another advantage of this change is that the same code can be shared
with UndoTupleSatisfiesUpdate, which also needs CID visibility
checks.