ZHeapTupleSatisfies: Improve logic for deferring CID lookup.
Instead of ZHeapTupleSatisfies guess at the correct value of
fetch_cid, and then have ZHeapTupleGetTransInfo sometimes ignore
it if it thinks that the caller guessed wrong, defer CID lookup
to the point when we discover that we need it to make a
visibility decision, and then go get it if we don't have it
already.

This doesn't quite get rid of the need for the fetch_cid
parameter to ZHeapTupleGetTransInfo, because it's still used
by ZHeapTupleSatisfiesUpdate.  But it at least gets rid of
the need for the snapshot parameter, since ZHeapTupleSatsisfies
was the only caller using that parameter, and now it doesn't.