Bug fix in applying undo action
When undo action is applied from the backend there is a small window
where the discard worker can push the request to rollback hash table before
the backend actually apply undo action.  Now, if backend apply the undo
actions and rewinds the insert location then the worker might read undo
records of the next transaction.  So for avoiding this backend need to
remove the entry from the hash table if it presents.

Dilip Kumar bug raised by Robert Haas