Don't store ZHeapTupleData fields in update/delete undo records.
Doing so is redundant, because the tuple length is already stored in
uur_tuple.len, the TID is already stored in uur_block and uur_offset,
and the table OID is already stored in uur_reloid.  Just get the
information from there instead of storing it all again.

Not only does this save on I/O, but it's also less code.