Prevent wrong WAL ordering of discard and Update TransInfo.
Both the discard and the update TransInfo logic register WAL but the
updaters only wait on the discard lock while updating the oldest_data.
This sometimes cause the update WAL to be written before the discard
WAL, causing error at REDO because it will expect the buffers
to be registered for the data considered discarded during DO time.

To fix this add a new lock discard_update_lock which will be used in
shared mode by the updaters and in exclusive mode during the actual
discard process.

Patch by me, reviewed by Dilip Kumar