Store undo record length at the end of the undo record.
Instead of kepping the track of the previous undo record in the
undo log meta and later updating same in the next undo record
directly store the undo record length at the end of each undo
record.  This will make code simpler and also for recovery we
don't need to log the prevlen in WAL as we can directly read
last 2 byte to read the record.

Dilip Kumar Idea by Thomas Munro and Reviewed by Amit Kapila