Perform undo actions for SPI rollbacks
For SPI rollbacks, we've to perform the undo actions. If the transaction
state is in progress, we perform the undo action before aborting the
transaction. Else, we save the undo record pointers so that it can be
rolled back later.

Patch by me. Reviewed by Amit Kapila and Dilip Kumar. Test case added by Neha
Sharma.