Refactor ZHeapTupleSatisfies to reduce code duplication.
The result of this is a bit clunky, because some of the more obscure
snapshot types require significantly different handling, but it avoids
a good bit of code duplication, because there are a bunch of common
steps that are performed for multiple snapshot types.  With more work,
this can probably be further streamlined.