Restructure fetching of tuples from undo.
GetTupleFromUndo now has the ability to determine whether a visible
version of the tuple exists without actually fetching the tuple from
the undo log; we do that only if the caller needs it.  Eventually,
this can be used by ZHeapTupleSatisfies, some of whose callers do not
need the returned tuple, but I haven't tried to make that work in this
commit.

The easiest way to implement this seemed to eliminate the last
remaining caller of CopyTupleFromUndoRecord and pull up the logic for
ZHeapTuple construction into GetTupleFromUndoRecord, so this commit
also removes CopyTupleFromUndoRecord.  It preserves some of the
commments in UpdateTupleHeaderFromUndoRecord which has assumed what
remains of CopyTupleFromUndoRecord's original function.

GetTupleFromUndo also now has the ability to test according to
satisfies-update rules; previously, it could only test according to
satisfies-MVCC and satisfies-self rules.  This eliminates the need for
a separate copy of the logic in UndoTupleSatisfiesUpdate, so that has
been removed in favor of having the callers invoke GetTupleFromUndo
directly.  These callers benefit from GetTupleFromUndo's new ability
to find out whether a visible version exists without necessarily
fetching it.