Remove GetTupleFromUndoForAbortedXact.
This code path is just fetching the latest committed version of the
tuple when the most recent update is known to have aborted.  But
GetTupleFromUndo can also tell us that.

GetTupleFromUndoForAbortedXact was probably marginally more efficient
than GetTupleFromUndo since it had various special-purpose
optimizations.  But those seem less important than not having multiple
copies of this code.  If they are important, perhaps we should add
some of them to GetTupleFromUndo so that all callers can benefit from
them.

For example, once we hit an undo record for an update or delete for a
committed XID, we should be able to just assume that the XID for any
previous update or delete is also committed.  That's because you can't
update or delete a tuple while another update or delete is in
progress.  Figuring out how to optimize this kind of thing might be
important for reducing CLOG pressure, or maybe it doesn't matter, but
we shouldn't have multiple copies of the code some of which try to
optimize this and others of which don't.

Unfortunately, I was not able to figure out a way to hit this
code, so add a ZBORKED comment saying it needs tests.