Tighten up code in ZHeapTupleSatisfiesUpdate.
Rather than having a bunch of separate call sites in this function
that each call GetTupleFromUndo and/or ZHeapTupleGetSubXid, use some
flags to indicate whether those operations are needed and have only
one place where we actually make the function call.

Also, use ZHeapTidOpFromInfomask for consistency with other places
where I made similar adjustments.

Also, don't bother catering to subxid == NULL, because no caller does
that.

This commit qualifies as a minor performance optimization and/or bug
fix, because in the old code arrangement if had an in-place updated or
ZHEAP_XID_LOCK_ONLY tuple, and then we went down the branch for
TransactionIdIsCurrentTransactionId, and then GetTupleFromUndo
returned false, we would not return.  Instead we'd fall through to the
code intended to handle newly-inserted tuples, which would recheck
TransactionIdIsCurrentTransactionId and then decide that the answer is
HeapTupleInvisible.  That's probably the right answer, but not the
right way to arrive at it.