Fixed truncation bug for non-index table
In lazy_scan_zheap, we were using tupindex varibale to
mark dead tuples to unused. In case of non-index table,
we were not reseting this variable for every block but
we were resting num_dead_tuples so we were not marking
dead tuples as unused hence we were not truncating.
So fixed this bug.

Patch by Mahendra Singh Thalor, reviewed by Amit Kapila,
reported by Hitachi