Remove ZHeapPageGetCid.
Use FetchTransInfoFromUndo instead.  It will overwrite zinfo.xid
and zinfo.epoch_xid, but it should find the same XID that was
already there; otherwise, the transaction slot point to some other
transaction's undo chain.  It will also update zinfo.urec_ptr,
but just to point to the record we actually found instead of the
record where we started the search, so it should be OK.

One perhaps-noteworthy difference is that FetchTransInfoFromUndo
skips UNDO_XID_LOCK_ONLY and UNDO_XID_MULTI_LOCK_ONLY records,
whereas ZHeapPageGetCid does not.  I *think* that should be a bug
fix if it makes any difference at all, because the point of
fetching the CID is to check visibility, and locking a tuple
can't change anything about the visibility of that tuple.