Add new interface function ZHeapTupleFetch and use it everywhere.
Previously, callers that were looked at non-deleted items used
ZHeapTupleSatisfies, while those looking at deleted items used
ZHeapGetVisibleTuple.  The new function ZHeapTupleFetch can handle
both cases, which in some places permits considerable simplification
on the calling side.  Currently, the new function just calls one or the
other of ZHeapTupleSatisfies and ZHeapGetVisibleTuple, but it should
be possible to combine more of the logic in the future.

The new function offers the option either to return the tuple or
just a true/false value.  A few callers now use this to avoid
leaking the returned tuple.  It also offers the option to return
the new CTID whether or not the tuple is deleted, where previously that
was only conveniently available for non-deleted tuple.  Those
operating on deleted tuples had to separately call
ZHeapPageGetNewCtid; this function will now do that automatically
where required, and in the future it should be possible to adjust
things so that we get the same effect with less redundant work.

The mysterious ValidateTuplesXact function now takes a relation as
an additional argument so that it can call ZHeapTupleFetch with
the right arguments.  There are other ways this could work, but
this seems simplest.

In various places, this converts the code to use zheap_gettuple
rather than bespoke logic for constructing a stack-allocated
ZHeapTupleData.  There might be a minor loss of efficiency there,
but I think it's worth it for the improved code cleanliness.
Further refactoring work will probably recapture any performance
we're surrendering here.