Consolidate and simplify code in ZHeapTupleFetch.
Most of the logic in ZHeapTupleSatifies moves into ZHeapTupleFetch,
with some modifications intended to make the logic work for deleted
tuples.

The role of ZHeapTupleSatisfies is reduced to determining whether a
tuple satisfies a snapshot, just like the name of the function might
suggest.

ZHeapGetVisibleTuple is removed, since the code paths for deleted
and non-deleted items have now been merged.  This has the side
effect that, if using a dirty snapshot on a deleted item, we will
now apply ZHeapSelectVersionDirty rather than ZHeapSelectVersionSelf.
That seems like a good thing, but will change the behavior when
TransactionIdIsCurrentTransactionId(xid) is false but
TransactionIdIsInProgress(xid) is true.  Also, SNAPSHOT_ANY will
now apply the appropriate set of rules rather than SnapshotSelf
rules.

I noticed that at some point during previous refactorings I inverted
the logic for the SNAPSHOT_NON_VACUUMABLE case, so flip that back
around.  It's surprising this doesn't break any tests.

The new logic for SNAPSHOT_NON_VACUUMABLE won't work in the case of
a tuple that's been deleted from the page, so Assert() that this
doesn't happen.  The old code would have applied SNAPSHOT_SELF
semantics, which doesn't seem like the right thing either.